Refactor utilities and move ornl_utils.py to pymbir.py.
