+++ mbirjax/preprocess/__init__.py
@@ -1,3 +1,5 @@
 from .utilities import *
 from .vcls import *
+from .segmentation import *
+from beam_hardening import *
 from . import nsi
