+++ mbirjax/preprocess/utilities.py
@@ -1,6 +1,5 @@
 import numpy as np
 import warnings
-import scipy
 import tifffile
 import glob
 import os
@@ -8,7 +7,6 @@ import jax.numpy as jnp
 import jax
 import dm_pix
 import tqdm
-
 
 
 def compute_sino_transmission(obj_scan, blank_scan, dark_scan, defective_pixel_array=(), batch_size=90):
@@ -483,331 +481,8 @@ def project_vector_to_vector(u1, u2):
     return u1_proj
 
 
-def multi_threshold_otsu(image, classes=2, num_bins=256):
-    """
-    Segment an image into multiple intensity classes using Otsu's method.
-
-    This function computes optimal threshold values that divide an image into the specified
-    number of classes by minimizing the intra-class variance. It returns `classes - 1` thresholds
-    that can be used to partition the image intensity range into `classes` distinct segments.
-
-    Args:
-        image (np.ndarray):
-            Input image as a NumPy array of floating-point values.
-        classes (int, optional):
-            Number of classes to divide the image into. Must be ≥ 2. Defaults to 2.
-        num_bins (int, optional):
-            Number of bins to use when constructing the image histogram. Defaults to 256.
-
-    Returns:
-        list of float:
-            A list of `classes - 1` threshold values, given in increasing order. These thresholds
-            can be used to separate the image into `classes` distinct intensity regions.
-
-    Example:
-        >>> thresholds = multi_threshold_otsu(image, classes=4)
-        >>> # Resulting thresholds will split image into 4 intensity regions
-    """
-    if classes < 2:
-        raise ValueError("Number of classes must be at least 2")
-
-    if num_bins < classes:
-        raise ValueError("Number of bins must be at least equal to number of classes")
-
-    # Compute the histogram of the image
-    hist, bin_edges = np.histogram(image, bins=num_bins, range=(np.min(image), np.max(image)))
-
-    # Find the optimal thresholds using a recursive approach
-    thresholds = _recursive_otsu(hist, classes - 1)
-
-    # Convert histogram bin indices to original image values
-    bin_centers = (bin_edges[:-1] + bin_edges[1:]) / 2
-    scaled_thresholds = [bin_centers[t] for t in thresholds]
-
-    return scaled_thresholds
-
-
-def _recursive_otsu(hist, num_thresholds):
-    """
-    Recursively applies Otsu's method to find the best thresholds for multiple classes.
-
-    Parameters
-    ----------
-    hist : ndarray
-        Histogram of the image.
-    num_thresholds : int
-        Number of thresholds to find.
-
-    Returns
-    -------
-    list
-        List of thresholds that divide the histogram into the specified number of classes.
-    """
-    # Base case: no thresholds needed
-    if num_thresholds == 0:
-        return []
-
-    # Base case: single threshold needed
-    if num_thresholds == 1:
-        return [_binary_threshold_otsu(hist)]
-
-    best_thresholds = []
-    best_variance = float('inf')
-
-    # Iterate through possible thresholds
-    for t in range(1, len(hist) - 1):
-        # Split histogram at the threshold
-        left_hist = hist[:t]
-        right_hist = hist[t:]
-
-        # Recursively find thresholds for left and right segments
-        left_thresholds = _recursive_otsu(left_hist, num_thresholds // 2)
-        right_thresholds = _recursive_otsu(right_hist, num_thresholds - len(left_thresholds) - 1)
-
-        # Combine thresholds
-        thresholds = left_thresholds + [t] + [x + t for x in right_thresholds]
-
-        # Compute the total within-class variance
-        total_variance = _compute_within_class_variance(hist, thresholds)
-
-        # Update the best thresholds if the current variance is lower
-        if total_variance < best_variance:
-            best_variance = total_variance
-            best_thresholds = thresholds
-
-    return best_thresholds
-
-
-def _binary_threshold_otsu(hist):
-    """
-    Finds the best threshold for binary segmentation using Otsu's method.
-
-    Parameters
-    ----------
-    hist : ndarray
-        Histogram of the image.
-
-    Returns
-    -------
-    int
-        Best threshold for binary segmentation.
-    """
-    total = np.sum(hist)
-    current_max, threshold = 0, 0
-    sum_total, sum_foreground, weight_foreground, weight_background = 0, 0, 0, 0
-
-    # Compute the sum of pixel values
-    for i in range(len(hist)):
-        sum_total += i * hist[i]
-
-    # Iterate through possible thresholds
-    for i in range(len(hist)):
-        weight_foreground += hist[i]
-        if weight_foreground == 0:
-            continue
-        weight_background = total - weight_foreground
-        if weight_background == 0:
-            break
-
-        sum_foreground += i * hist[i]
-        mean_foreground = sum_foreground / weight_foreground
-        mean_background = (sum_total - sum_foreground) / weight_background
-
-        # Compute between-class variance
-        between_class_variance = weight_foreground * weight_background * (mean_foreground - mean_background) ** 2
-        if between_class_variance > current_max:
-            current_max = between_class_variance
-            threshold = i
-
-    return threshold
-
-
-def _compute_within_class_variance(hist, thresholds):
-    """
-    Computes the total within-class variance given a set of thresholds.
-
-    Parameters
-    ----------
-    hist : ndarray
-        Histogram of the image.
-    thresholds : list
-        List of thresholds that divide the histogram into multiple classes.
-
-    Returns
-    -------
-    float
-        Total within-class variance.
-    """
-    total_variance = 0
-    thresholds = [0] + thresholds + [len(hist)]
-
-    # Iterate through each segment defined by the thresholds
-    for i in range(len(thresholds) - 1):
-        class_hist = hist[thresholds[i]:thresholds[i+1]]
-        class_prob = np.sum(class_hist)
-        if class_prob == 0:
-            continue
-        class_mean = np.sum(class_hist * np.arange(thresholds[i], thresholds[i+1])) / class_prob
-        class_variance = np.sum(((np.arange(thresholds[i], thresholds[i+1]) - class_mean) ** 2) * class_hist) / class_prob
-        total_variance += class_variance * class_prob
-
-    return total_variance
 # ####### END Multi-threshold Otsu's method
 
 
 
-# ####### Generalized Huber Weights method
-def gen_huber_weights(weights, sino_error, T=1.0, delta=1.0, epsilon=1e-6):
-    """
-    This function generates generalized Huber weights based on the method described in the referenced notes.
-    It adds robustness by treating any element where ``|sino_error / weights| > T`` as an outlier,
-    down-weighting it according to the generalized Huber function.
 
-    The function returns new `ghuber_weights`.
-
-    Typically, to obtain the final robust weights, the `ghuber_weights` should be multiplied by the original `weights`:
-
-        final_weights = weights * ghuber_weights
-
-    Args:
-        weights: jnp.ndarray of shape (views, rows, cols):
-            Initial weights, typically derived from inverse variance estimates.
-        sino_error: jnp.ndarray of shape (views, rows, cols):
-            Sinogram error array representing deviations from the model.
-        T: float, optional (default=1.0):
-            Threshold parameter; values greater than T are treated as outliers.
-        delta: float, optional (default=1.0):
-            Controls the strength of the generalized Huber function (delta=1 corresponds to the conventional Huber).
-        epsilon: float, optional (default=1e-6):
-            Small number to avoid division by zero.
-
-    Returns:
-        huber_weights: jnp.ndarray of shape (views, rows, cols)
-            The computed generalized Huber weights.
-
-    Notes:
-        The generalized Huber function used in this function is based on:
-        Venkatakrishnan, S. V., Drummy, L. F., Jackson, M., De Graef, M., Simmons, J. P., and Bouman, C. A.,
-        "Model-Based Iterative Reconstruction for Bright-Field Electron Tomography,"
-        IEEE Transactions on Computational Imaging, vol. 1, no. 1, pp. 1–15, 2015. DOI: 10.1109/TCI.2014.2371751
-
-    Example:
-        >>> from mbirjax import gen_huber_weights
-        >>> huber_weights = gen_huber_weights(weights, sino_error)
-        >>> final_weights = weights * huber_weights
-    """
-    if not (0.0 <= delta <= 1.0):
-        raise ValueError("delta must be between 0 and 1.")
-
-    weights = jnp.asarray(weights)
-    sino_error = jnp.asarray(sino_error)
-
-    # Compute std and global alpha
-    std = 1.0 / jnp.maximum(jnp.sqrt(weights), epsilon)
-    alpha = jnp.linalg.norm(sino_error) / (jnp.linalg.norm(std) + epsilon)
-    std_norm = alpha * std
-
-    # Compute normalized error
-    normalized_error = sino_error / std_norm
-    abs_norm_error = jnp.abs(normalized_error)
-
-    # Apply generalized Huber function
-    huber_weights = jnp.where(abs_norm_error <= T, 1.0, (delta * T) / (abs_norm_error + epsilon))
-
-    return huber_weights
-
-def BH_correction(sino, alpha, batch_size=64):
-    """
-    Apply a polynomial beam hardening correction to a sinogram.
-
-    This function applies a polynomial correction to each view of the sinogram
-    by evaluating powers of the sinogram values and weighting them by the coefficients in `alpha`,
-    while also including the original linear term (the sinogram itself).
-
-    The corrected sinogram is computed as:
-
-        corrected_sino = sino + alpha[0] * sino**2 + alpha[1] * sino**3 + ...
-
-    It processes the sinogram in batches of views for memory efficiency.
-
-    Args:
-        sino (jnp.ndarray or np.ndarray of shape (views, rows, cols)):
-            Input sinogram to correct.
-        alpha (list or array of floats):
-            Coefficients for the polynomial correction. The k-th term corresponds to sino^(k+1).
-        batch_size (int, optional, default=16):
-            Number of views to process in a single batch.
-
-    Returns:
-        corrected_sino: jnp.ndarray of shape (views, rows, cols)
-            Beam hardening corrected sinogram.
-
-    Example:
-        >>> from mbirjax.preprocess import BH_correction
-        >>> alpha = [1.0, 0.2, 0.1]  # Correction: sino + 0.2 * sino^2 + 0.1 * sino^3
-        >>> corrected_sino = BH_correction(sino, alpha)
-    """
-    # Ensure inputs are JAX arrays
-    sino = jnp.asarray(sino)
-    alpha = jnp.asarray(alpha)
-
-    views, rows, cols = sino.shape
-    corrected = []
-
-    for i in range(0, views, batch_size):
-        sino_batch = sino[i:i+batch_size]
-
-        # Initialize corrected batch to the linear term (sino_batch)
-        corrected_batch = jnp.array(sino_batch)
-
-        # Apply polynomial terms
-        for k in range(len(alpha)):
-            corrected_batch += alpha[k] * jnp.power(sino_batch, k + 1)
-
-        corrected.append(corrected_batch)
-
-    corrected_sino = jnp.concatenate(corrected, axis=0)
-
-    return corrected_sino
-
-
-def segment_plastic_metal(recon):
-    """
-    Segment a reconstruction into plastic and metal masks using multi-threshold Otsu.
-
-    This function uses multi-threshold Otsu segmentation to classify the input 
-    reconstruction into several classes and returns binary masks for the plastic 
-    and metal components. It also returns scaling factor representing the average value of the recon in the segmented region.
-    
-    Args:
-        recon (jnp.ndarray): Reconstructed volume array.
-
-    Returns:
-        Tuple[jnp.ndarray, jnp.ndarray, float, float]: A tuple containing:
-            - plastic_mask (jnp.ndarray): Binary mask for plastic regions.
-            - metal_mask (jnp.ndarray): Binary mask for metal regions.
-            - plastic_scale (float): Scaling factor for the plastic mask.
-            - metal_scale (float): Scaling factor for the metal mask.
-
-    Example:
-        >>> import mbirjax as mj
-        >>> import mbirjax.preprocess as mjp
-        >>> plastic_mask, metal_mask, plastic_scale, metal_scale = mjp.segment_plastic_metal(recon)
-        >>> mj.slice_viewer(plastic_mask, metal_mask, vmin=0, vmax=1.0, 
-        ...                 slice_label=['Plastic', 'Metal'], 
-        ...                 title='Plastic and Metal Masks')
-    """
-    # Determine class thresholds based on the 3-classes
-    thresholds = multi_threshold_otsu(recon, classes=3)
-    plastic_low_threshold = thresholds[0]
-    plastic_metal_threshold = thresholds[1]
-
-    # Create masks
-    plastic_mask = jnp.where((recon > plastic_low_threshold) & (recon <= plastic_metal_threshold), 1.0, 0.0)
-    metal_mask = jnp.where(recon > plastic_metal_threshold, 1.0, 0.0)
-
-    # Scale factors that match the unitary masks to the reconstruction
-    plastic_scale = _compute_scaling_factor(recon, plastic_mask)
-    metal_scale = _compute_scaling_factor(recon, metal_mask)
-
-    return plastic_mask, metal_mask, plastic_scale, metal_scale
